requirements.txt
easydict==1.7
tensorflow-gpu==1.11
scipy==0.18.1
numpy==1.15.4
opencv_python==3.4.0.12
Cython==0.27.3
Pillow==5.0.0
PyYAML==3.12
tensorflow-serving-api==1.11
keras==2.2.4
